backend/requirements.txt
fastapi==0.111.0
aerich
tortoise-orm
pytesseract
pdf2image
numpy
opencv